Added exercises 1.32 and 1.33.
These exercises include refactoring some procedures, such as 'sum' and
'product'. Thus, I also changed some file names to reflect the
newfound procedure definitions within.

Future git tip to think about: in the previous commit I staged and
committed the removal 1.3's readme file (because it was in the wrong
place and I moved it to where it was supposed to be), but I did not
stage the change of place of this readme file. I avoided this because
I already made changes to the readme file, but as a result that
commit simply doesn't have 1.3's readme anywhere.

What I should've done instead was to stash the changes I've made, then
staged and committed the new location of the readme, then reapply
the stashed changes. That way the readme still would've existed in the
previous commit. Oh well, you live you learn.