Added exercises 2.5 through 2.9.
Have started working the interval arithmetic extended exercise, and as
such I have created and provided modules for the differing
abstraction barriers of the interval arithmetic package.
`intervals.rkt` is concerned only with interval
representation/implementation (rather easy as it is simply a
`cons`, a `car`, and a `cdr` so far). `interval-arithmetic.rkt` is,
of cource, concerned with using the interval abstractions to perform
arithmetic with them.