Added exercises 2.28 and 2.29.
Ideally, I would have created two different modules for exercise 2.29 -
one that is only concerned with the representation of the mobiles, and
one that uses that representation to define its own procedures. This
would enforce the abstraction barrier a little more. However, I got lazy
and didn't want to do all of that, so they are in the same racket
module.