Added exercises 2.68 through 2.72.
Finished section 2.3! Huffman trees were a bit harder to understand,
though with some thinking I was able to wrap my head around it. Proud of
myself for this one.

The use of symbolic data is interesting, and it further allows us to
abstract certain details away as we no longer have to work explicity
with numerical values. On to the next!