Began working on data-directed programming in chapter 2.3.
Also created the necessary module for being able to manipulate an
operator lookup table (or a hash table) that is used rather extensively
in this chapter. The book goes under the assumption that we have `get`
and `put` as primitives, however, in order to test out my solutions I
needed these implemented.

The procedures provided in the hash-table.rkt module now allow me to
"install" certain packages within the lookup table, as well as be able
to dispatch on type using type-tagging.