Finally got the arithmetic package to work. Will be using the sicp language and Dr. Racket in order to work with mutable cells/pairs.
This took a while - one for figuring out a solution to making these
mutable pairs work, and another because I was in the middle of a move!
Ayayay, but hopefully this is all set up now and things are smooth
sailing from here on out. Will need to monitor any side effects of using
the sicp/Scheme language, and for using the Dr. Racket IDE.