Made the apply-generic procedure more general in order to handle coercion with multiple arguments.
Does so by first attempting to coerce all elements of the argument list
into the same type as the first element. If that does not work, try the
second element, and so on.

Within each arithmetic package, I also included coercion procedures that
outline how to coerce one element to another. This perhaps should not be
included in each arithmetic package themselves, and should probably be
part of a separate package that can be installed - this would align
better with the principle of modularity and additivity. However, since
this is part of my system that I've developed, I've decided that that
location makes the most sense for the exercise problems, so that way I
can keep like with like.