Added exercise 2.83, which introduces the generic `raise` procedure.
We broke up the scheme number package into two separate packages:
integers and reals. We then introduced a method for raising a
lower-level data type to a higher-level one - in this instance it goes
like this:
integer -> rational -> real -> complex

Looking ahead to the next couple of exercises, we will have to modify
our `apply-generic` procedure in our table operations module in order to
use `raise`, which is defined in our arithmetic package module. This
means that I will have to merge the two files into one in order to
prevent any messy coupling or circular dependencies.

Womp womp.