Exercise 2.85. Implemented our simplifying drop procedure.
To achieve this, we also implemented a project procedure, which attempts
to push the data type down one level in the type hierarchy. The drop
procedure then checks to see if the projected number is equivalent to
the original number when we raise the projected number again. If it is,
then we can continue attempting to drop the data object.